Enhancement Implemented
1.Dealer/Sub Dealer/Support Executive > Enrollment >At the time of customer enrolment from, mobile number validation needs to be added.
2.Sub Dealer/Engineer/Mason > My purchase claim > Order Quantity needs to be replaced with Trade Quantity
3.Sub Dealer/Engineer/Mason > Purchase Request > a. During the claim process in the mobile app, no need to display the selection of user type.
b. By default the mapped dealer and subdealer will be listed in the dropdown along with the firm name. Customer can search for the delaer/sub dealer name by name or firm name.
4.Sub Dealer/Engineer/Mason > Cash Transfer > a. At the time of cash transfer no need to display the selection of user type.
b. By default all the mapped dealer and sub delaer will be listed in the dropdown, customer can select the dealer / sub delaer and request for the cash transfer.
5.Dealer  > Dashboard > Dealer stock inventory (Add product stock availability list).
6. Engineer/Mason > Worksite  > a. Auto location capture and time
b. For engineer, Hide the Engineer details field